refactor(grasping): respect isolated runtime project boundaries
