+++ dimos/manipulation/grasping/grasp_gen_x/test_module.py
@@ -12,39 +12,29 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Hermetic tests for the import-safe GraspGenX adapter."""
+"""Host contract tests; the isolated runtime is not imported here."""
 
 from __future__ import annotations
 
 import inspect
+from pathlib import Path
 import subprocess
 import sys
 from typing import Any
 
 import numpy as np
 import pytest
-from pytest_mock import MockerFixture
 
-from dimos.experimental.isolated_python.bootstrap import validate_runtime
 from dimos.manipulation.grasping.grasp_gen_spec import GraspGenSpec, LegacyGraspGenSpec
-from dimos.manipulation.grasping.grasp_gen_x.module import (
-    GraspGenXConfig,
-    GraspGenXError,
-    GraspGenXModule,
-)
-from dimos.manipulation.grasping.grasp_gen_x.python.graspgenx_runtime import runtime as grasp_gen_x
-from dimos.manipulation.grasping.grasp_gen_x.python.graspgenx_runtime.runtime import (
-    _GraspGenXRuntimeModule,
-)
+from dimos.manipulation.grasping.grasp_gen_x.module import GraspGenXConfig
 from dimos.msgs.geometry_msgs.Pose import Pose
 from dimos.msgs.manipulation_msgs.GraspCandidate import GraspCandidate
 from dimos.msgs.manipulation_msgs.GraspCandidateArray import GraspCandidateArray
-from dimos.msgs.sensor_msgs.PointCloud2 import PointCloud2
 from dimos.msgs.std_msgs.Header import Header
 
 
-def config(**overrides: object) -> GraspGenXConfig:
-    values: dict[str, object] = {
+def config(**overrides: Any) -> GraspGenXConfig:
+    values: dict[str, Any] = {
         "gripper": {
             "extents_open": (0.1, 0.1, 0.1),
             "offset_open": (0.0, 0.0, 0.0),
@@ -55,15 +45,6 @@ def config(**overrides: object) -> GraspGenXConfig:
     }
     values.update(overrides)
     return GraspGenXConfig(**values)  # type: ignore[arg-type]
-
-
-def module_args(value: GraspGenXConfig | None = None) -> dict[str, Any]:
-    return (value or config()).model_dump(exclude={"rpc_transport", "tf_transport", "g"})
-
-
-def cloud(points: np.ndarray | None = None) -> PointCloud2:
-    xyz = np.zeros((1, 3), dtype=np.float32) if points is None else points
-    return PointCloud2.from_numpy(xyz, frame_id="camera", timestamp=12.5)
 
 
 def test_public_adapter_import_does_not_load_optional_runtime() -> None:
@@ -83,17 +64,6 @@ def test_public_adapter_import_does_not_load_optional_runtime() -> None:
     )
 
     assert result.returncode == 0, result.stderr
-
-
-@pytest.fixture
-def runtime(mocker: MockerFixture) -> Any:
-    create_runtime = mocker.patch.object(grasp_gen_x, "_create_runtime")
-    instance = create_runtime.return_value
-    instance.infer.return_value = (
-        np.repeat(np.eye(4, dtype=np.float32)[None], 1, axis=0),
-        np.asarray([0.5], dtype=np.float32),
-    )
-    return create_runtime
 
 
 def test_messages_round_trip_empty_and_score() -> None:
@@ -137,7 +107,7 @@ def test_ranked_spec_is_canonical_during_legacy_contract_transition() -> None:
         ("fingertip_depth", 0.0),
     ],
 )
-def test_gripper_constraints_are_declared_by_fields(field: str, value: object) -> None:
+def test_gripper_constraints_are_declared_by_fields(field: str, value: Any) -> None:
     gripper = config().gripper.model_dump()
 
     with pytest.raises(ValueError):
@@ -145,7 +115,7 @@ def test_gripper_constraints_are_declared_by_fields(field: str, value: object) -
 
 
 @pytest.mark.parametrize("value", [0, -1, True])
-def test_candidate_limit_is_a_strict_positive_integer(value: object) -> None:
+def test_candidate_limit_is_a_strict_positive_integer(value: Any) -> None:
     with pytest.raises(ValueError):
         config(max_candidates=value)
 
@@ -162,145 +132,22 @@ def test_rigid_transform_relational_validation() -> None:
         )
 
 
-def test_start_is_synchronous_and_idempotent(runtime: Any) -> None:
-    module = _GraspGenXRuntimeModule(_isolated_python_runtime=True, **module_args())
-    try:
-        module.start()
-        module.start()
-
-        runtime.assert_called_once_with(module.config)
-        assert len(module.propose_grasps(cloud())) == 1
-    finally:
-        module.stop()
-
-
-def test_start_failure_is_explicit(runtime: Any) -> None:
-    runtime.side_effect = RuntimeError("CUDA unavailable")
-    module = _GraspGenXRuntimeModule(_isolated_python_runtime=True, **module_args())
-    try:
-        with pytest.raises(GraspGenXError, match="initialize"):
-            module.start()
-    finally:
-        module.stop()
-
-
-def test_adapter_sorts_stably_truncates_and_applies_tcp_transform(runtime: Any) -> None:
-    poses = np.repeat(np.eye(4, dtype=np.float32)[None], 3, axis=0)
-    poses[:, 0, 3] = [1.0, 2.0, 3.0]
-    runtime.return_value.infer.return_value = (
-        poses,
-        np.asarray([0.5, 0.5, 0.9], dtype=np.float32),
+def test_host_collection_excludes_the_nested_runtime_suite() -> None:
+    result = subprocess.run(
+        [
+            sys.executable,
+            "-m",
+            "pytest",
+            "--collect-only",
+            "--no-cov",
+            "-q",
+            str(Path(__file__).parent),
+        ],
+        capture_output=True,
+        text=True,
+        timeout=60,
     )
-    cfg = config(
-        max_candidates=2,
-        grasp_frame_to_tcp=(
-            (0.0, -1.0, 0.0, 10.0),
-            (1.0, 0.0, 0.0, 0.0),
-            (0.0, 0.0, 1.0, 0.0),
-            (0.0, 0.0, 0.0, 1.0),
-        ),
-    )
-    module = _GraspGenXRuntimeModule(_isolated_python_runtime=True, **module_args(cfg))
-    try:
-        module.start()
-        result = module.propose_grasps(cloud())
 
-        assert [candidate.score for candidate in result] == pytest.approx([0.9, 0.5])
-        assert [candidate.pose.position.x for candidate in result] == pytest.approx([13.0, 11.0])
-        assert result.header.frame_id == "camera"
-        assert result.header.timestamp == pytest.approx(12.5)
-    finally:
-        module.stop()
-
-
-def test_empty_backend_result_preserves_input_header(runtime: Any) -> None:
-    runtime.return_value.infer.return_value = (
-        np.empty((0, 4, 4), dtype=np.float32),
-        np.empty(0, dtype=np.float32),
-    )
-    module = _GraspGenXRuntimeModule(_isolated_python_runtime=True, **module_args())
-    try:
-        module.start()
-        result = module.propose_grasps(cloud())
-
-        assert result.header.frame_id == "camera"
-        assert result.header.timestamp == pytest.approx(12.5)
-        assert result.candidates == []
-    finally:
-        module.stop()
-
-
-@pytest.mark.parametrize(
-    "points",
-    [
-        np.array([[np.nan, 0.0, 0.0]], dtype=np.float32),
-        np.empty((0, 3), dtype=np.float32),
-        np.zeros((2, 2), dtype=np.float32),
-    ],
-)
-def test_invalid_cloud_points_are_rejected(runtime: Any, points: np.ndarray) -> None:
-    module = _GraspGenXRuntimeModule(_isolated_python_runtime=True, **module_args())
-    try:
-        module.start()
-        with pytest.raises(ValueError, match="pointcloud|XYZ"):
-            module.propose_grasps(cloud(points))
-        runtime.return_value.infer.assert_not_called()
-    finally:
-        module.stop()
-
-
-@pytest.mark.parametrize(
-    "backend",
-    [
-        (np.ones((2, 4, 4)), np.ones(1)),
-        (np.full((1, 4, 4), np.nan), np.ones(1)),
-        (np.ones((1, 4, 4)), np.array([np.inf])),
-    ],
-)
-def test_invalid_backend_outputs_are_rejected(
-    runtime: Any, backend: tuple[np.ndarray, np.ndarray]
-) -> None:
-    runtime.return_value.infer.return_value = backend
-    module = _GraspGenXRuntimeModule(_isolated_python_runtime=True, **module_args())
-    try:
-        module.start()
-        with pytest.raises(ValueError):
-            module.propose_grasps(cloud())
-    finally:
-        module.stop()
-
-
-def test_inference_failure_is_wrapped(runtime: Any) -> None:
-    runtime.return_value.infer.side_effect = RuntimeError("backend")
-    module = _GraspGenXRuntimeModule(_isolated_python_runtime=True, **module_args())
-    try:
-        module.start()
-        with pytest.raises(GraspGenXError, match="inference"):
-            module.propose_grasps(cloud())
-    finally:
-        module.stop()
-
-
-def test_not_started_and_missing_metadata_are_rejected(runtime: Any) -> None:
-    module = _GraspGenXRuntimeModule(_isolated_python_runtime=True, **module_args())
-    missing_frame = cloud()
-    missing_frame.frame_id = ""
-    missing_timestamp = cloud()
-    missing_timestamp.ts = None
-    try:
-        with pytest.raises(GraspGenXError, match="not been started"):
-            module.propose_grasps(cloud())
-        module.start()
-        with pytest.raises(ValueError, match="frame_id"):
-            module.propose_grasps(missing_frame)
-        with pytest.raises(ValueError, match="timestamp"):
-            module.propose_grasps(missing_timestamp)
-    finally:
-        module.stop()
-
-
-def test_runtime_implements_the_host_contract() -> None:
-    validate_runtime(GraspGenXModule, _GraspGenXRuntimeModule)
-    assert inspect.signature(GraspGenXModule.propose_grasps) == inspect.signature(
-        _GraspGenXRuntimeModule.propose_grasps
-    )
+    assert result.returncode == 0, result.stdout + result.stderr
+    assert "test_public_adapter_import_does_not_load_optional_runtime" in result.stdout
+    assert "test_runtime.py" not in result.stdout
