+++ dimos/robot/test_all_blueprints_generation.py
@@ -116,9 +116,7 @@ def _build_module_class_set(root: Path) -> set[str]:
     known: set[str] = {"Module", "ModuleBase"}
     all_classes: list[tuple[str, list[str]]] = []
 
-    for path in sorted(root.rglob("*.py")):
-        if "__pycache__" in str(path):
-            continue
+    for path in sorted(_get_all_python_files(root)):
         try:
             tree = ast.parse(path.read_text("utf-8"), str(path))
         except Exception:
@@ -256,11 +254,17 @@ def _check_for_uncommitted_changes(file_path: Path) -> bool:
 
 
 def _get_all_python_files(root: Path) -> Generator[Path, None, None]:
-    for path in root.rglob("*.py"):
-        rel_path = str(path.relative_to(root.parent))
-        if "__pycache__" in str(path) or rel_path in IGNORED_FILES:
-            continue
-        yield path
+    for directory, children, filenames in os.walk(root):
+        parent = Path(directory)
+        children[:] = [
+            name
+            for name in children
+            if name != "__pycache__" and not (parent / name / "pyproject.toml").is_file()
+        ]
+        for name in filenames:
+            path = parent / name
+            if path.suffix == ".py" and str(path.relative_to(root.parent)) not in IGNORED_FILES:
+                yield path
 
 
 def _path_to_module_name(path: Path, root: Path) -> str:
@@ -325,3 +329,26 @@ def _ends_with_blueprint_method(node: ast.expr) -> bool:
         if isinstance(func, ast.Attribute) and func.attr in BLUEPRINT_METHODS:
             return True
     return False
+
+
+def test_nested_projects_do_not_contribute_modules_or_blueprints(tmp_path: Path) -> None:
+    root = tmp_path / "dimos"
+    runtime = root / "provider/python"
+    runtime.mkdir(parents=True)
+    (runtime / "pyproject.toml").write_text('[project]\nname = "runtime"\n')
+    (root / "provider/contract.py").write_text(
+        "class HostContract(Module): pass\n"
+        "class RuntimeOnlyBase: pass\n"
+        "class NotAModule(RuntimeOnlyBase): pass\n"
+        "host_blueprint = HostContract.blueprint()\n"
+    )
+    (runtime / "runtime.py").write_text(
+        "class PublicRuntime(HostContract): pass\n"
+        "class RuntimeOnlyBase(Module): pass\n"
+        "runtime_blueprint = PublicRuntime.blueprint()\n"
+    )
+
+    blueprints, modules = _scan_for_blueprints(root)
+
+    assert blueprints == {"host-blueprint": "dimos.provider.contract:host_blueprint"}
+    assert modules == {"host-contract": "dimos.provider.contract.HostContract"}
